requirements.txt
python-dotenv
kaggle
tensorflow
matplotlib
numpy
ipython
nbconvert
jupyter